+++ src/libs/lighter/src/Window.cpp
@@ -1248,11 +1248,11 @@ void Window::LoadPreset(long prs)
         switch (list[i].type)
         {
         case ListElement::t_smooth:
-            smoothNorm = ini->GetLong(sect, "smoothUseNormals", static_cast<long>(smoothNorm)) != 0;
+            smoothNorm = ini->GetInt(sect, "smoothUseNormals", static_cast<long>(smoothNorm)) != 0;
             smoothRad = static_cast<float>(ini->GetDouble(sect, "smoothRadius", smoothRad));
             break;
         case ListElement::t_blur:
-            isTraceBlur = ini->GetLong(sect, "blurTrace", static_cast<long>(isTraceBlur)) != 0;
+            isTraceBlur = ini->GetInt(sect, "blurTrace", static_cast<long>(isTraceBlur)) != 0;
             blurRad = static_cast<float>(ini->GetDouble(sect, "blurRadius", blurRad));
             blurAtt = static_cast<float>(ini->GetDouble(sect, "blurAtt", blurAtt));
             blurCos = static_cast<float>(ini->GetDouble(sect, "blurCos", blurCos));
@@ -1280,7 +1280,7 @@ void Window::LoadPreset(long prs)
             *list[i].gamma =
                 static_cast<float>(ini->GetDouble(sect, GenerateName(list[i].name, "_gamma"), *list[i].gamma));
             *list[i].isOn =
-                ini->GetLong(sect, GenerateName(list[i].name, "_isOn"), static_cast<long>(*list[i].isOn)) != 0;
+                ini->GetInt(sect, GenerateName(list[i].name, "_isOn"), static_cast<long>(*list[i].isOn)) != 0;
             break;
         case ListElement::t_glight:
             list[i].st = static_cast<float>(ini->GetDouble(sect, GenerateName(list[i].name, "_intens"), list[i].st));
@@ -1306,7 +1306,7 @@ void Window::LoadPreset(long prs)
             *list[i].range =
                 static_cast<float>(ini->GetDouble(sect, GenerateName(list[i].name, "_range"), *list[i].range));
             *list[i].isOn =
-                ini->GetLong(sect, GenerateName(list[i].name, "_isOn"), static_cast<long>(*list[i].isOn)) != 0;
+                ini->GetInt(sect, GenerateName(list[i].name, "_isOn"), static_cast<long>(*list[i].isOn)) != 0;
             break;
         }
     }
