+++ src/libs/weather/src/Weather.h
@@ -42,7 +42,7 @@ class WEATHER : public WEATHER_BASE
     void CleanUP();
 
     // Weather_Base function
-    long GetLong(uint32_t dwCode) override;
+    long GetInt(uint32_t dwCode) override;
     uint32_t GetColor(uint32_t, CVECTOR *) override;
     uint32_t GetColor(uint32_t) override;
     float GetFloat(uint32_t) override;
