+++ src/libs/xinterface/src/Nodes/xi_scrollimage.cpp
@@ -6,7 +6,7 @@ long GetTexFromEvent(VDATA *vdat)
 {
     if (vdat == nullptr)
         return -1;
-    return vdat->GetLong();
+    return vdat->GetInt();
 }
 
 CXI_SCROLLIMAGE::CXI_SCROLLIMAGE()
